bytemuck_util: rename and flatten `pod`
Rename the `pod` submodule to `anybitpattern`, since that is the type it
now implements.

Also "flatten" by making this submodule private, and re-exporting at the
top level of the crate, as `bytemuck` does. e.g.
`bytemuck_util::AnyBitPattern` instead of
`bytemuck_util::anybitpattern::AnyBitPattern`.
