+++ src/lib/bytemuck-util/src/lib.rs
@@ -1,1 +1,6 @@
-pub mod pod;
+mod anybitpattern;
+
+pub use anybitpattern::{
+    maybeuninit_bytes_of, maybeuninit_bytes_of_mut, maybeuninit_bytes_of_slice,
+    maybeuninit_bytes_of_slice_mut, zeroed, AnyBitPattern,
+};
