+++ src/main/host/descriptor/socket/inet/legacy_tcp.rs
@@ -3,7 +3,6 @@ use std::net::{Ipv4Addr, SocketAddrV4};
 use std::sync::Arc;
 
 use atomic_refcell::AtomicRefCell;
-use bytemuck_util::pod;
 use nix::errno::Errno;
 use nix::sys::socket::{MsgFlags, Shutdown, SockaddrIn};
 use shadow_shim_helper_rs::syscall_types::ForeignPtr;
@@ -820,7 +819,7 @@ impl LegacyTcpSocket {
             return Err(Errno::EINVAL.into());
         }
 
-        let mut peer_addr: libc::sockaddr_in = pod::zeroed();
+        let mut peer_addr: libc::sockaddr_in = bytemuck_util::zeroed();
         peer_addr.sin_family = libc::AF_INET as u16;
         let mut accepted_fd = -1;
 
@@ -925,7 +924,7 @@ impl LegacyTcpSocket {
     ) -> Result<libc::socklen_t, SyscallError> {
         match (level, optname) {
             (libc::SOL_TCP, libc::TCP_INFO) => {
-                let mut info = pod::zeroed();
+                let mut info = bytemuck_util::zeroed();
                 unsafe { c::tcp_getInfo(self.as_legacy_tcp(), &mut info) };
 
                 let optval_ptr = optval_ptr.cast::<crate::cshadow::tcp_info>();
