+++ src/main/host/syscall/io.rs
@@ -1,7 +1,6 @@
 use std::mem::MaybeUninit;
 use std::ops::{Deref, DerefMut};
 
-use bytemuck_util::pod;
 use nix::errno::Errno;
 use shadow_shim_helper_rs::syscall_types::ForeignPtr;
 
@@ -140,7 +139,7 @@ pub fn read_sockaddr(
 /// # Ok(())
 /// # }
 /// ```
-pub fn write_partial<T: pod::AnyBitPattern>(
+pub fn write_partial<T: bytemuck_util::AnyBitPattern>(
     mem: &mut MemoryManager,
     val: &T,
     val_ptr: ForeignPtr<T>,
@@ -148,7 +147,7 @@ pub fn write_partial<T: pod::AnyBitPattern>(
 ) -> Result<usize, Errno> {
     let val_len_bytes = std::cmp::min(val_len_bytes, std::mem::size_of_val(val));
 
-    let val = &pod::maybeuninit_bytes_of(val)[..val_len_bytes];
+    let val = &bytemuck_util::maybeuninit_bytes_of(val)[..val_len_bytes];
 
     let val_ptr = val_ptr.cast::<MaybeUninit<u8>>();
     let val_ptr = ForeignArrayPtr::new(val_ptr, val_len_bytes);
